+++ lib/config.ml
@@ -15,18 +15,11 @@
  *)
 open Import
 
-let base_packages =
-  [
-    "jbuilder";
-    "dune";
-    "ocamlbuild";
-    "ocamlmod";
-    "oasis";
-    "ocamlify";
-    "ocaml";
-    "ocaml-base-compiler";
-    "ocaml-variants";
-  ]
+(* These packages are usually used only for legacy building. Since
+   opam-monorepo requires everything to build with dune (which is special
+   cased) we can ignore these packages if they're in the dependency formula *)
+let skip_packages =
+  [ "jbuilder"; "dune"; "ocamlbuild"; "ocamlmod"; "oasis"; "ocamlify" ]
   |> List.map ~f:OpamPackage.Name.of_string
   |> OpamPackage.Name.Set.of_list