+++ lib/opam.mli
@@ -20,15 +20,17 @@ module Package_summary : sig
     hashes : OpamHash.t list;
     dev_repo : string option;
     depexts : (OpamSysPkg.Set.t * OpamTypes.filter) list;
+    flags : OpamTypes.package_flag list;
   }
 
   val pp : t Fmt.t
   val from_opam : OpamPackage.t -> OpamFile.OPAM.t -> t
 
-  val is_virtual : t -> bool
-  (** A package is considered virtual if it has no url.src or no dev-repo. *)
+  val is_safe_package : t -> bool
+  (** A package is safe when it is clear that it can be added to the lockfile
+      without disruption, as it will not get pulled.
 
-  val is_base_package : t -> bool
+      The OCaml compiler is one such package. *)
 end
 
 module Dependency_entry : sig